fix(build): above-anchor y is a Float chain — coerceAtLeast(0f).toInt() before IntOffset

+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/menu/LyricsMenu.kt
@@ -2355,7 +2355,8 @@ fun AnchoredLyricsOverflowMenu(
                                 // below) lands right on the icon.
                                 (iconBoundsInRoot.top - verticalOffsetPx -
                                     (if (popupHeightPx > 0) popupHeightPx else with(density) { 360.dp.toPx() }.toInt()))
-                                    .coerceAtLeast(0)
+                                    .coerceAtLeast(0f)
+                                    .toInt()
                             } else {
                                 iconBottom + verticalOffsetPx
                             }
